% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFscore.R
\name{CFscore}
\alias{CFscore}
\title{Counterfactual validation score}
\usage{
CFscore(
  object,
  data,
  outcome,
  treatment_formula,
  treatment_of_interest,
  metrics = c("auc", "brier", "oeratio", "calplot"),
  time_horizon,
  cens_model = "KM",
  cens_formula = ~1,
  null_model = TRUE,
  stable_iptw = FALSE,
  bootstrap = 0,
  bootstrap_progress = TRUE,
  iptw,
  ipcw,
  quiet = FALSE
)
}
\arguments{
\item{object}{One of the following three options to be validated:
\itemize{
  \item a numeric vector, corresponding to risk predictions
  \item a glm or coxph model
  \item a (named) list, with one or more of the previous 2 options, for
  validating and comparing multiple models at once.
}}

\item{data}{A data.frame containing the observed outcome, assigned treatment,
and necessary confounders for the validation of object.}

\item{outcome}{The outcome, to be evaluated within data. This could either be
the name of a numeric column in data, or a Surv object for time-to-event
data, e.g. Surv(time, status), if time and status are columns in data.}

\item{treatment_formula}{A formula which identifies the treatment (left hand
side) and the confounders (right hand side) in the data. E.g. A ~ L. The
confounders are used to estimate the inverse probability of treatment
weights (IPTW) model. The IPTW can also be specified themselves using the
iptw argument, in which case the right hand side of this formula is
ignored.}

\item{treatment_of_interest}{A treatment level for which the counterfactual
perormance measures should be evaluated.}

\item{metrics}{A character vector specifying which performance metrics to be
computed. Options are c("auc", "brier", "oeratio", "calplot"). See details.}

\item{time_horizon}{For time to event data, the prediction horizon of
interest.}

\item{cens_model}{Model for estimating inverse probability of censoring
weights (IPCW). Methods currently implemented are Kaplan-Meier ("KM") or
Cox ("cox") both applied to the censored times.}

\item{cens_formula}{Formula for which the r.h.s. determines the censoring
probabilities. I.e. ~ x1 + x2.}

\item{null_model}{If TRUE fit a risk prediction model which ignores the
covariates and predicts the same value for all subjects. The model is
fitted using the data in which all subjects are counterfactually assigned
the treatment of interest (using the IPTW, as estimated using the
treatment_formula or as given by the iptw argument). For time-to-event
outcomes, the subjects are also counterfactually uncensored (using the
IPCW, as estimated using the outcome_formula, or as given by the ipcw
argument).}

\item{stable_iptw}{if TRUE, estimate stabilized IPTW weights. See details.}

\item{bootstrap}{If this is an integer greater than 0, this indicates the
number of bootstrap iterations, to compute 95\% confidence intervals around
the performance metrics.}

\item{bootstrap_progress}{if set to TRUE, print a progress bar indicating the
progress of bootstrap procedure.}

\item{iptw}{A numeric vector, containing the inverse probability of treatment
weights. These are normally computed using the treatment_formula, but they
can be specified directly via this argument. If specified via this
argument, bootstrap is not possible.}

\item{ipcw}{A numeric vector, containing the inverse probability of censor
weights. These are normally computed using the outcome_formula, but they
can be specified directly via this argument. If specified via this
argument, bootstrap is not possible.}

\item{quiet}{If set to TRUE, don't print assumptions.}
}
\value{
A list with performance metrics.
}
\description{
Counterfactual validation score
}
\details{
auc is area under the (ROC) curve, estimated ...

  Brier score is defined as 1 / sum(iptw) sum(predictions_i - outcome_i)^2
  scaled brier score is also possible (metric = "scaled_brier")

  oeratio represents the observed/expected ratio, where observed is the mean
  of the outcomes in the pseudopopulation. The expected is the mean of the
  predictions in the original observed population.
}
\examples{
n <- 1000

data <- data.frame(L = rnorm(n), P = rnorm(n))
data$A <- rbinom(n, 1, plogis(data$L))
data$Y <- rbinom(n, 1, plogis(0.1 + 0.5*data$L + 0.7*data$P - 2*data$A))

random <- runif(n, 0, 1)
model <- glm(Y ~ A + P, data = data, family = "binomial")
naive_perfect <- data$Y

CFscore(
  object = list("ran" = random, "mod" = model, "per" = naive_perfect),
  data = data,
  outcome = Y,
  treatment_formula = A ~ L,
  treatment_of_interest = 0,
)
}
