% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_score_long.R
\name{ip_score_long}
\alias{ip_score_long}
\title{Interventional prediction score for longitudinal treatment strategies}
\usage{
ip_score_long(
  predictions,
  data_outcome,
  data_long,
  treatment_formula,
  treatment_of_interest,
  metrics = c("auc", "brier", "scaled_brier", "oeratio", "calplot"),
  visit_times,
  time_horizon,
  cens_model = "KM",
  cens_formula = ~1,
  null_model = TRUE,
  bootstrap = 0,
  bootstrap_progress = TRUE,
  iptw,
  ipcw,
  quiet = FALSE,
  strip_ipt_models = TRUE
)
}
\arguments{
\item{predictions}{A numeric vector corresponding to the risk estimates
under evaluation, or a (named) list of multiple numeric vectors for
evaluating and comparing multiple vectors.}

\item{data_outcome}{A dataframe containing the observed outcomes. It must
consist of 1 row per subject, with columns 'id', 'time', and 'status'. Time
can be continuous, status represents the binary outcome status at ‘time’.}

\item{data_long}{A dataframe in 'long' format, i.e., each subject has one row for each of their visits. Contains the time-dependent
treatment variable, the (potentially time-dependent) adjustment variables
(confounders) and other time dependent covariates that may be required for
modeling the censoring mechanism. Baseline covariates can be included by repeating their value for every visit. Must
include the columns 'id' and 'visit_time'. There should not be any visits
after a subject's follow-up time. All subjects should have data at each
visit.}

\item{treatment_formula}{A formula which indicates the treatment/intervention
(left hand side) and the adjustment variables (right hand side), which must be present in
data_long. E.g. A ~ L + A_lag_1. The left hand side can be either a binary
treatment (coded as 0/1 numeric, logical or factor) or a treatment with more
than two categories (coded as factor). The adjustment variables (right hand
side) are used to estimate the inverse probability of treatment weights
(IPTW) with logistic/multinomial regression. The IPTW can also be specified directly using the iptw argument, in which
case the right hand side of this formula is ignored (the left hand side must
still indicate the treatment, i.e. A ~ 1).}

\item{treatment_of_interest}{A treatment strategy under which the predictions
should be evaluated. Should be in the form of a vector, with one element per
visit. See details.}

\item{metrics}{A character vector specifying which performance metrics to be
computed. Options are c("auc", "brier", “scaled_brier”, "oeratio",
"calplot").}

\item{visit_times}{A numeric vector, indicating the times of the visits at
which sequential treatment decisions were made. The first visit must always
be at time 0. Visit times are not necessarily equidistant.}

\item{time_horizon}{the prediction horizon of interest. Should be after the
last visit.}

\item{cens_model}{Model for estimating inverse probability of censored weights
(IPCW). Methods currently implemented are Kaplan-Meier ("KM") or Cox
("cox"), with censoring times derived from the (time,status) variables in
data_outcome, reversing the status indicator, see details. KM is only
supported when the right hand side of cens_formula is 1.}

\item{cens_formula}{Model formula used for estimating the censoring
probabilities. Could consist of only baseline variables but
also of time-dependent variables. All variables specified must be present in
data_long. The left hand side must be left blank. E.g. ~ x1 + x2.}

\item{null_model}{If TRUE fits a model without covariates that estimates the same probability for all subjects in data. The model is
fitted using the reweighted data in which all subjects 'counterfactually'
received the treatment level of interest (using the IPTW, as estimated
using the treatment_formula or as given by the iptw argument). For
time-to-event outcomes, the null model is also fitted using the IPCW, as estimated using the cens_formula, or as given by the ipcw argument. The null_model can be used as reference (baseline) model.}

\item{bootstrap}{If this is an integer greater than 0, this indicates the
number of bootstrap iterations, to compute 95\% confidence intervals around
the performance metrics based on percentiles of the bootstrap results.}

\item{bootstrap_progress}{if set to TRUE, print a progress bar indicating the
progress of bootstrap procedure.}

\item{iptw}{A numeric vector with length nrow(data_outcome), containing one
over the probability of being compliant to the treatment strategy of interest
for each subject. If iptw is not specified, these weights are computed using the treatment_formula, but they can be specified
directly via this argument. A user-defined function can also be specified,
which takes as input arguments 'data_outcome' and 'data_long' and returns a
numeric vector of the IPTW weights. See details.}

\item{ipcw}{A numeric vector with length nrow(data_outcome), containing the inverse probability of censoring
weights at the time_horizon, or at a subject's event time, whichever happens
first. For subjects who are censored before the time_horizon, the ipcw can be left at NA. If ipcw is not specified, these weights are computed using the
cens_formula, but they can be specified
directly via this argument. A user-defined function can also be specified,
which takes as input arguments 'data_outcome' and 'data_long' and returns a
numeric vector of the IPCW weights. See details.}

\item{quiet}{If set to TRUE, don't print assumptions.}

\item{strip_ipt_models}{If set to TRUE (default), unnecessary components from
the IPT- and IPC-model objects are not stored to save memory. Set to FALSE
if you want to store the full IPT/IPC model objects.}
}
\value{
An object of class `ip_score`, for which the `print()` and `plot()`
methods are implemented. The object is a nested list containing: \itemize{
  \item `$score`, contains the estimated predictive performance metrics.
  \item `$bootstrap`, if requested, the 95\% confidence intervals of the
  performance metrics, and the performance metrics for each individual
  bootstrap iteration.
  \item `$outcome`, the observed outcomes from data_outcome.
  \item `$treatment`, the observed treatment levels from data_long.
  \item `$predictions`, the predictions to be evaluated, i.e. the estimated probability
  of event under the intervention strategy of interest for each subject.
  \item `$ipt`, method, model and inverse probability of treatment weights
  (IPTW). The IPTW are NA for subjects who did not follow the treatment strategy of interest.
  \item `$ipc`, method, model and inverse probability of censoring weights
  (IPCW). The IPCW are NA for subjects who were censored before the time_horizon.
  \item `$pseudopop`, binary vector indicating which subjects in data_outcome were re-weighted to create the pseudo-population. These are the subjects who followed the
  treatment strategy of interest and were not censored before the time_horizon, if applicable.
  }
 The print method summarizes the results and if (quiet = FALSE), prints the
 assumptions required for valid inference.
}
\description{
Estimates the predictive performance of predictions of binary or
time-to-event outcomes under longitudinal intervention strategies, by
reweighting the data to form a pseudo-population in which every subject was
assigned the longitudinal treatment strategy of interest.
}
\details{
To form a pseudo-population that represents a setting in which everybody
received treatment level 1 during five visits, set `treatment_of_interest` to
`c(1,1,1,1,1)`. Any pattern can be chosen, e.g. `c(1,0,1,0,1)` is also valid,
as long as the number of visits equals ‘n_visits’. Alternatively, one could
also set this argument to "always" or "never" as a shortcut for `rep(1,
n_visits)` or `rep(0, n_visits)`. Treatment strategies where treatment is only
set at certain visits are also possible via `NA`, e.g. use `c(1,1,NA, NA, NA)`
when you want to form a pseudo-population where everybody's treatment levels
are set to 1 at the first two visits, and their treatment level at the remaining three visits can be whatever
they would have normally been (natural course) after those first two.
If treatment is categorical, `treatment_of_interest` should be a character vector, e.g `c(“active”,”control”).

All performance metrics are computed on the weighted population mimicking the
hypothetical situation where every subject’s treatment strategy was set to the
treatment strategy of interest (and where nobody was censored before time_horizon). "auc" is area
under the (ROC) curve. "brier" is Brier score, ranging from 0 to 1. Scaled
brier score is also available (metrics = "scaled_brier"), which expresses the Brier score relative to the null model. For the O/E ratio,
the numerator (observed) is the (weighted) fraction of 'observed' events in
the pseudo-population, and the denominator (expected) is the (unweighted) mean
of risk estimates for all subjects in data. The calplot option generates a
calibration plot, with default 8 subgroups. More/less subgroups can be
specified by appending “calplot” with a number indicating the number of
subgroups, e.g. metrics = "calplot10" for 10 subgroups.

The null model estimates are obtained as the weighted mean outcome in the subset
that followed the treatment strategy of interest (and was not censored before
time_horizon). For time-to-event data, this null prediction could also be
computed using a weighted Kaplan-Meier estimator, which would be more
efficient, but computationally slower.

The censoring distribution is estimated with a Kaplan-Meier estimator implemented using `prodlim::prodlim(...,
reverse = TRUE)`. This correctly estimates the censoring distribution when
there are ties between event and censoring times. When using a Cox model to
estimate the censoring distribution, the event indicator is reversed manually. This does
not preserve the usual tie-handling convention: in standard survival analysis,
censoring is assumed to occur after events at the same time point, but after
reversing the indicator the opposite ordering is assumed. A possible
workaround is to add a small positive offset (`epsilon`) to all censoring
times before fitting the censoring model.

When supplying time-to-event data in which some subjects have follow-up
beyond the specified time horizon, the full follow-up time is used when
estimating the censoring distribution, rather than truncating follow-up at
the prediction horizon. This does not affect the IPCW when the censoring
distribution is estimated using the Kaplan-Meier estimator, but it can
affect the IPCW when using a Cox model. If the censoring model should be
estimated using information only up to the prediction horizon, the user must
administratively censor subjects at the prediction horizon before supplying
the data to `ip_score_long()`.

Bootstrapping is not possible when manually specifiying the IPTW/IPCW as
numeric vectors. If specifying a model user-defined function that computes the
ITPW/IPCW given data, it is possible. The given function will be called on
each bootstrapped dataset and resulting metrics are used to compute the
95\% CIs with the percentile method. More advanced techniques such as thresholding extreme IP
weights, can be implemented through a user-defined weight function. The
censoring weight returned by this function should be the 1 / probability of
remaining uncensored at the time horizon, or at their event time, whichever
happens first. For subjects who are censored before the time_horizon, the ipcw can be left at NA.
}
\examples{
set.seed(5)
n <- 1000
data <- data.frame(id = 1:n)

# 2 visits at t = 0 and t = 2. Time dependent confounding between A and L
# Here we generate random survival times between 0 and 6 as an example.
# See generating longitudinal data vignette for another example where survival outcomes
# depend on time dependent  L and A.
data <- within(data, {
  L0 <- rnorm(n)
  A0 <- rbinom(n, 1, plogis(0.7*L0))
  L1 <- rnorm(n, 0.8*L0 - 0.2*A0)
  A1 <- rbinom(n, 1, plogis(0.7*L1 + 0.6*A0))
  time <- runif(n, 0, 6)
  status <- rbinom(n, 1, 0.5)
})
# A0 is at time t = 0, A1 is at time t = 2

# If you have data in this 'wide' form, then usually you can use the convenience
# functions to make it suitable for ip_score_long
head(data)
# note that due to our simulation there may be measurements after a subject was
# censored or had event. This will be removed later.

data_outcome <- data[, c("id", "time", "status")]
data_long <- wide_to_long(
  df = data,
  baseline_variables = "id",
  wide_variables = list("A" = c("A0", "A1"),
                        "L" = c("L0", "L1")),
  visit_times = c(0,2),
  outcome_times = data$time
)
# adding treatment level at previous visit time to the data. For the first
# visit this is assumed to be 0 (fill)
data_long <- add_lag_terms(data_long, var = "A", lag = 1, fill = 0)

head(data_long)
# note that the measurements that were generated after follow up time have been
# removed by wide_to_long().

# To get some predictions to evaluate, here is a model that
# randomly predicts a number between 0.2 and 0.8
random_model <- runif(n, min = 0.2, max = 0.8)

# Our treatment at visit i depends on confounder value L at same visit and,
# if visit = 2, on treatment level at previous visit. We will illustrate a correctly
# specified IPTW model with A ~ L + A_lag_1.

# Estimate performance of random predictions in a pseudopopulation in
# which everybody was assigned treatment strategy {0, 0}
ip_score_long(
  predictions = random_model,
  data_outcome = data_outcome,
  data_long = data_long,
  treatment_formula = A ~ L + A_lag_1,
  treatment_of_interest = c(0,0),
  visit_times = c(0,2),
  time_horizon = 4
)

# Performance of random predictions in a pseudopopulation in
# which everybody was assigned treatment level 1 at visit 1, and whatever
# they would normally have after that (natural course) at visit 2, and modelling
# the censoring distribution with Cox model
ip_score_long(
  predictions = random_model,
  data_outcome = data_outcome,
  data_long = data_long,
  treatment_formula = A ~ L + A_lag_1,
  treatment_of_interest = c(1,NA),
  visit_times = c(0,2),
  time_horizon = 4,
  cens_model = "cox",
  cens_formula = ~ A + A_lag_1 + L
)
}
\references{
Keogh RH, Van Geloven N. Prediction Under Interventions:
 Evaluation of Counterfactual Performance Using Longitudinal Observational
 Data. Epidemiology. 2024;35(3):329-339.
}
\seealso{
\code{\link{ip_score}}

Open the corresponding vignette for more extensive examples with
\code{vignette("longitudinal", package = "ipeval")}
}
