% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_score.R
\name{ip_score}
\alias{ip_score}
\title{Counterfactual validation score}
\usage{
ip_score(
  object,
  data,
  outcome,
  treatment_formula,
  treatment_of_interest,
  metrics = c("auc", "brier", "oeratio", "calplot"),
  time_horizon,
  cens_model = "KM",
  cens_formula = ~1,
  null_model = TRUE,
  stable_iptw = FALSE,
  bootstrap = 0,
  bootstrap_progress = TRUE,
  iptw,
  ipcw,
  quiet = FALSE
)
}
\arguments{
\item{object}{One of the following three options to be validated:
\itemize{
  \item a numeric vector, corresponding to risk predictions under
  intervention of interest.
  \item a glm or coxph model, capable of estimating risks under intervention
  of interest. See details.
  \item a (named) list, with one or more of the previous 2 options, for
  validating and comparing multiple models at once.
}}

\item{data}{A data.frame containing the observed outcome, assigned treatment,
and necessary confounders for the validation of object.}

\item{outcome}{The outcome, to be evaluated within data. This could either be
the name of a numeric/logical column in data, or a Surv object for
time-to-event data, e.g. Surv(time, status), if time and status are columns
in data.}

\item{treatment_formula}{A formula which identifies the treatment (left hand
side) and the confounders (right hand side) in the data. E.g. A ~ L. The
confounders are used to estimate the inverse probability of treatment
weights (IPTW) model. The IPTW can also be specified themselves using the
iptw argument, in which case the right hand side of this formula is
ignored.}

\item{treatment_of_interest}{A treatment level for which the counterfactual
perormance measures should be evaluated.}

\item{metrics}{A character vector specifying which performance metrics to be
computed. Options are c("auc", "brier", "oeratio", "calplot"). See details.}

\item{time_horizon}{For time to event data, the prediction horizon of
interest.}

\item{cens_model}{Model for estimating inverse probability of censored
weights (IPCW). Methods currently implemented are Kaplan-Meier ("KM") or
Cox ("cox"), both applied to the censored times. KM is only supported when
the right hand side of cens_formula is 1.}

\item{cens_formula}{Formula for which the r.h.s. determines the censoring
probabilities. I.e. ~ x1 + x2.}

\item{null_model}{If TRUE fit a risk prediction model which ignores the
covariates and predicts the same value for all subjects. The model is
fitted using the data in which all subjects are counterfactually assigned
the treatment of interest (using the IPTW, as estimated using the
treatment_formula or as given by the iptw argument). For time-to-event
outcomes, the subjects are also counterfactually uncensored (using the
IPCW, as estimated using the cens_formula, or as given by the ipcw
argument).}

\item{stable_iptw}{if TRUE, estimate stabilized IPT-weights. See details.}

\item{bootstrap}{If this is an integer greater than 0, this indicates the
number of bootstrap iterations, to compute 95\% confidence intervals around
the performance metrics.}

\item{bootstrap_progress}{if set to TRUE, print a progress bar indicating the
progress of bootstrap procedure.}

\item{iptw}{A numeric vector, containing the inverse probability of treatment
weights. These are normally computed using the treatment_formula, but they
can be specified directly via this argument. If specified via this
argument, bootstrap is not possible.}

\item{ipcw}{A numeric vector, containing the inverse probability of censor
weights. These are normally computed using the cens_formula, but they can
be specified directly via this argument. If specified via this argument,
bootstrap is not possible.}

\item{quiet}{If set to TRUE, don't print assumptions.}
}
\value{
An object of class `ip_score`, for which the `print()` and `plot()`
methods are implemented. The object is a nested list containing: \itemize{
  \item `$score`, which contains the predictive performance in the 'counterfactual'
  dataset.
  \item `$outcome`, the observed outcome of the original dataset.
  \item `$treatment`, the observed outcome of the original dataset.
  \item `$predictions`, the predictions to be evaluated, i.e. the probability
  of event for each patient, had their treatment been set to
  treatment_of_interest
  \item `$ipt`, method, model and inverse probability of treatment weights
  (IPTW). These are NA for patients that are not in the pseudopopulation.
  \item `$ipc`, method, model and inverse probability of censoring weights (IPCW).
     these are NA for patients that were censored.
  \item `$correct_trt`, binary vector indicating which subjects of the original
     population followed the treatment of interest.
  \item `$uncensored`, binary vector indicating which subjects of the original
     population were uncensored, if applicable.
  \item `$pseudopop`, binary vector indicating which subjects of the original
     population were in the pseudopopulation. Equal to
     `$correct_trt & $uncensored`.
  }
}
\description{
Estimates the predictive performance of predictions under interventions, by
forming a weighted pseudopopulation in which every subject was assigned the
treatment of interest.
}
\details{
When supplying a glm or coxph model, the model should be able to estimate
risks under the intervention of interest. This could be done in two ways: the
model does not have a treatment covariate, and always estimates the risk
under intervention of interest. Alternatively, the model has a covariate for
treatment. This function then automatically estimates the risk under the
treatment of interest for all subjects, even if they were assigned
alternative treatment.

All performance metrics are computed on the weighted population in which
every subject was counterfactually assigned treatment of interest. \code{auc}
is area under the (ROC) curve. Brier score is defined as \code{1 / sum(iptw)
sum(predictions_i - outcome_i)^2}. Scaled brier score is also available
(\code{metrics = "scaled_brier"}). For the O/E ratio, the numerator
(observed) is the weighted fraction of events in the pseudopopulation, and
the denominator (expected) is the unweighted mean of risk estimates of the
original unweighted population. The \code{calplot} option generates a
calibration plot, with default 8 knots. More/less knots can be specified by
appending calplot with a number indicating the number of knots, e.g.
\code{metrics = calplot10} for 10 knots.

For the null model, the O/E ratio and the scaled Brier score, the mean
predicted risk under the treatment of interest is required. This is computed
by the weighted mean in the ('counterfactually' uncensored) pseudopopulation.
For survival data, this null prediction could theoretically also be computed
with a weighted Kaplan-Meier, which is supposed to be more efficient, but
computationally a lot slower. Both methods are valid.

Stabilized IPT-weigths are computed by estimating a null model for treatment.
E.g. weights are \code{P(A = a) / P(A = a | L = l)}, if the given
treatment_formula is \code{A ~ L}.
}
\examples{
n <- 1000

data <- data.frame(L = rnorm(n), P = rnorm(n))
data$A <- rbinom(n, 1, plogis(data$L))
data$Y <- rbinom(n, 1, plogis(0.1 + 0.5*data$L + 0.7*data$P - 2*data$A))

random <- runif(n, 0, 1)
model <- glm(Y ~ A + P, data = data, family = "binomial")
naive_perfect <- data$Y

score <- ip_score(
  object = list("ran" = random, "mod" = model, "per" = naive_perfect),
  data = data,
  outcome = Y,
  treatment_formula = A ~ L,
  treatment_of_interest = 0,
)
print(score)
plot(score)
}
\references{
Keogh RH, Van Geloven N. Prediction Under Interventions:
  Evaluation of Counterfactual Performance Using Longitudinal Observational
  Data. Epidemiology. 2024;35(3):329-339.

  Boyer CB, Dahabreh IJ, Steingrimsson JA. Estimating and Evaluating
  Counterfactual Prediction Models. Statistics in Medicine.
  2025;44(23-24):e70287.

  Pajouheshnia R, Peelen LM, Moons KGM, Reitsma JB, Groenwold RHH. Accounting
  for treatment use when validating a prognostic model: a simulation study.
  BMC Medical Research Methodology. 2017;17(1):103.
}
