% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeval-package.R
\docType{package}
\name{ipeval-package}
\alias{ipeval}
\alias{ipeval-package}
\title{ipeval: Interventional Prediction Evaluation}
\description{
Provides methods to evaluate predictive performance of models that estimate risks under hypothetical intervention scenarios (interventional/causal/counterfactual predictions) with observational data subject to treatment-outcome confounding. Inverse probability of treatment weighting (IPTW) is used to construct a pseudopopulation in which all individuals receive a specified intervention, enabling assessment of agreement between predicted risks under the intervention and observed outcomes in the pseudo-population corresponding to that intervention. Supports interventions with binary or categorical treatment levels, applied either at a single time point or as longitudinal treatment strategies with sequential treatment decisions. Performance measures supported are AUC (Area Under the receiving operating characteristic Curve), Brier score, observed-expected ratio, and calibration plots. Methods implemented in this package are based on work by Keogh and Van Geloven (2024) \doi{10.1097/EDE.0000000000001713}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/survival-lumc/ipeval}
  \item \url{https://survival-lumc.github.io/ipeval/}
  \item Report bugs at \url{https://github.com/survival-lumc/ipeval/issues}
}

}
\author{
\strong{Maintainer}: Jasper van Egeraat \email{j.w.a.van_egeraat@lumc.nl}

Authors:
\itemize{
  \item Nan van Geloven [copyright holder]
  \item Ruth Keogh [copyright holder]
}

Other contributors:
\itemize{
  \item Leiden University Medical Center [funder]
}

}
\keyword{internal}
