% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_score.R
\name{observed_score}
\alias{observed_score}
\title{Performance in observed dataset This function exists only to demonstrate the
difference between 'normal' performance and counterfactual performance. It is
not user friendly and should not be used.}
\usage{
observed_score(
  object,
  data,
  outcome,
  metrics = c("auc", "brier", "oeratio", "calplot")
)
}
\value{
Performance metrics in the observed dataset.
}
\description{
Performance in observed dataset This function exists only to demonstrate the
difference between 'normal' performance and counterfactual performance. It is
not user friendly and should not be used.
}
