% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_score.R
\name{observed_score}
\alias{observed_score}
\title{Performance in observed dataset}
\usage{
observed_score(
  object,
  data,
  outcome,
  metrics = c("auc", "brier", "oeratio", "calplot")
)
}
\arguments{
\item{object}{One of the following three options to be validated:
\itemize{
  \item a numeric vector, corresponding to risk predictions
  \item a glm model
  \item a (named) list, with one or more of the previous 2 options, for
  validating and comparing multiple models at once.
}}

\item{data}{A data.frame containing the observed outcome.}

\item{outcome}{The outcome, to be evaluated within data. This should be
the name of a numeric column in data.}

\item{metrics}{A character vector specifying which performance metrics to be
computed. Options are c("auc", "brier", "oeratio", "calplot").}
}
\value{
Performance metrics in the observed dataset.
}
\description{
This function computes the performance of the predictions in the given data,
which may contain a mix of treated and untreated subjects. It exists only to
demonstrate the difference between 'normal' performance and counterfactual
performance. It is not user friendly and should not be relied on. It does not
support time-to-event data.
}
\examples{
n <- 1000

data <- data.frame(L = rnorm(n), P = rnorm(n))
data$A <- rbinom(n, 1, plogis(data$L))
data$Y <- rbinom(n, 1, plogis(0.1 + 0.5*data$L + 0.7*data$P - 2*data$A))

random <- runif(n, 0, 1)
model <- glm(Y ~ A + P, data = data, family = "binomial")

observed_score(
  object = list("ran" = random, "mod" = model),
  data = data,
  outcome = Y,
  metrics = c("auc", "brier", "oeratio")
)
}
